Add Show FAQ button in app header instead of Overview page
