+++ app/views/athletes/_header.html.erb
@@ -75,6 +75,11 @@
             <i class="fa fa-share-alt"></i>
           </a>
         </li>
+        <li>
+          <a class="show-faq" href="#" title="Show FAQ">
+            <i class="fa fa-question-circle"></i>
+          </a>
+        </li>
         <% if @is_current_user %>
           <li>
             <a href="#" class="control-sidebar-toggle" data-toggle="control-sidebar" title="Show Settings">
