+++ test/controllers/expected/athletes/9123806.html
@@ -103,6 +103,11 @@
             <i class="fa fa-share-alt"></i>
           </a>
         </li>
+        <li>
+          <a class="show-faq" href="#" title="Show FAQ">
+            <i class="fa fa-question-circle"></i>
+          </a>
+        </li>
           <li>
             <a href="https://www.strava.com/oauth/authorize?client_id=&amp;response_type=code&amp;redirect_uri=http://www.example.com:80/auth/exchange_token&amp;approval_prompt=auto&amp;scope=view_private" class="btn-connect-with-strava" title="Connect With Strava">
               <img alt="Connect With Strava" src="/assets/btn_strava_connectwith_orange.png" />
