+++ test/controllers/expected/athletes/9123806_logged_in.html
@@ -103,6 +103,11 @@
             <i class="fa fa-share-alt"></i>
           </a>
         </li>
+        <li>
+          <a class="show-faq" href="#" title="Show FAQ">
+            <i class="fa fa-question-circle"></i>
+          </a>
+        </li>
           <li>
             <a href="#" class="control-sidebar-toggle" data-toggle="control-sidebar" title="Show Settings">
               <i class="fa fa-gears"></i>
